[compiler] Fixing bugs in variance of function types
Variance for function types was outof-sync with the VM. At runtime, we do not allow to be `|&mut T|` to be assignable to `|&T|`. However, the compiler allowed this."

Fixing this problem was a bit more complicated by inline functions. For inline functions, historically the above is allowed, and we can't breal this. Thus the compiler needs to know in which context function types need to be checked.

This is implemented by adding a new `Variance::ShallowImplInlineVariance` to type resolution.

While debugging this, I also figured that the error messages are pretty bad around function types. This is improved in this PR as well, leading to a larger diff.